+++ src/components/cellTypes/FormulaCell.tsx
@@ -47,7 +47,11 @@ const FormulaCell = (mdProps: CellComponentProps) => {
     <span
       ref={formulaRef}
       className={`${c(
-        "md_cell " + getAlignmentClassname(tableColumn.config.content_alignment)
+        "md_cell " +
+          getAlignmentClassname(
+            tableColumn.config,
+            configInfo.getLocalSettings()
+          )
       )}`}
       key={`formula_${cell.id}`}
     />